move util test files to the test folder
