+++ src/app/views/query-runner/query-input/auto-complete/AutoComplete.styles.ts
