+++ src/app/views/query-runner/query-input/auto-complete/auto-complete.util.spec.ts
@@ -1,8 +1,9 @@
+import { IAutoCompleteProps } from '../../../../../types/auto-complete';
 import {
-  cleanUpSelectedSuggestion, getParametersWithVerb, getLastCharacterOf, getLastSymbolInUrl,
-  getFilteredSuggestions
-} from '../../app/views/query-runner/query-input/auto-complete/auto-complete.util';
-import { IAutoCompleteProps } from '../../types/auto-complete';
+  cleanUpSelectedSuggestion,
+  getParametersWithVerb, getLastCharacterOf,
+  getFilteredSuggestions, getLastSymbolInUrl
+} from './auto-complete.util';
 
 describe('Tests autocomplete utils', () => {
   it('Tests cleanUpSelectedSuggestion', () => {
@@ -78,3 +79,26 @@ describe('Tests autocomplete utils', () => {
     expect(result).toEqual({ key: '/', value: 32 });
   })
 })
+
+describe('Query input util should', () => {
+
+  it('cleanup selected suggestion', async () => {
+    const compareString = 'sel';
+    const userInput = 'https://graph.microsoft.com/v1.0/me/messages?sel';
+    const selected = '$select';
+    const selectedSuggestion = cleanUpSelectedSuggestion(compareString, userInput, selected);
+
+    expect(selectedSuggestion).toEqual('https://graph.microsoft.com/v1.0/me/messages?$select');
+  });
+
+  it('replace only last occurrence of compare string', async () => {
+    const compareString = 'su';
+    const userInput = 'https://graph.microsoft.com/v1.0/me/messages?$select=id,subject&orderby=su';
+    const selected = 'subject desc';
+    const selectedSuggestion = cleanUpSelectedSuggestion(compareString, userInput, selected);
+
+    expect(selectedSuggestion)
+      .toEqual('https://graph.microsoft.com/v1.0/me/messages?$select=id,subject&orderby=subject desc');
+  });
+
+});
