+++ src/app/views/sidebar/resource-explorer/panels/postman.util.spec.ts
@@ -1,10 +1,8 @@
-import content from '../../app/utils/resources/resources.json';
-import { generatePostmanCollection } from '../../app/views/sidebar/resource-explorer/panels/postman.util';
-import {
-  createResourcesList,
-  getResourcePaths
-} from '../../app/views/sidebar/resource-explorer/resource-explorer.utils';
-import { IResource } from '../../types/resources';
+import { IResource } from '../../../../../types/resources';
+import content from '../../../../utils/resources/resources.json';
+import { createResourcesList, getResourcePaths } from '../resource-explorer.utils';
+import { generatePostmanCollection } from './postman.util';
+
 const resource = JSON.parse(JSON.stringify(content)) as IResource;
 
 describe('Postman collection should', () => {
