+++ src/themes/theme-utils.spec.ts
@@ -1,4 +1,5 @@
-import { saveTheme, readTheme } from '../../../src/themes/theme-utils';
+import { saveTheme, readTheme } from './theme-utils';
+
 describe('Tests theme utils', () => {
   it('Saves theme to local storage then retrieves the saved theme', () => {
     const theme = 'dark';
